requirements.txt
python-dotenv
azure-common
azure-mgmt-resource
azure-mgmt-datafactory